Add null check when checking capabilities, move GetSharesForFileAsyncTask to ui.asynctasks package and other fixes
